# docs/openapi.yaml
openapi: 3.1.0
info:
  title: Hotel AI OS Core API
  version: 0.2.3-pilot.6
  description: >-
    PRODUCT-V1.2 / TECH-V0.2 Sprint 2.1 UAT-candidate contract. Production requests use a
    trusted JWT/SSO bearer token; roles, permissions, organization scope and
    active position assignments are resolved by the server. X-Tenant-Id and
    X-Actor-Id are accepted only when DEV_HEADER_AUTH_ENABLED=true for local
    integration. Client-supplied role or organization-scope headers are never
    trusted.
servers:
  - url: http://localhost:8080
security:
  - BearerAuth: []
  - TenantHeader: []
    ActorHeader: []
...
paths:
  /api/v1/org/units:
    get:
      tags: [Organization]
      operationId: listOrganizationUnits
      summary: List visible organization units
      parameters:
        - name: type
          in: query
          schema: { type: string, enum: [GROUP, REGION, HOTEL, DEPARTMENT] }
      responses:
        '200': { $ref: '#/components/responses/ListResponse' }
        '403': { $ref: '#/components/responses/Forbidden' }
    post:
      tags: [Organization]
      operationId: createOrganizationUnit
      summary: Create an organization unit
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/CreateOrgUnit' }
      responses:
        '201': { $ref: '#/components/responses/ObjectResponse' }
        '403': { $ref: '#/components/responses/Forbidden' }
  /api/v1/org/positions:
    get:
      tags: [Organization]
      operationId: listPositions
      summary: List position definitions
      responses:
        '200': { $ref: '#/components/responses/ListResponse' }
    post:
      tags: [Organization]
      operationId: createPosition
      summary: Create a position definition
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/CreatePosition' }
      responses:
        '201': { $ref: '#/components/responses/ObjectResponse' }
  /api/v1/org/units/{orgUnitId}:
    put:
      tags: [Organization]
      operationId: updateOrganizationUnit
      summary: Edit or activate/deactivate an organization unit without moving its hierarchy
      parameters:
        - { name: orgUnitId, in: path, required: true, schema: { type: string, format: uuid } }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/UpdateOrgUnit' }
      responses:
        '200': { $ref: '#/components/responses/ObjectResponse' }
        '403': { $ref: '#/components/responses/Forbidden' }
    delete:
      tags: [Organization]
      operationId: deleteUnusedOrganizationUnit
      summary: Delete an inactive, childless and unreferenced organization unit
      parameters:
        - { name: orgUnitId, in: path, required: true, schema: { type: string, format: uuid } }
      responses:
        '204': { description: Deleted }
        '403': { $ref: '#/components/responses/Forbidden' }
  /api/v1/org/positions/{positionId}:
    put:
      tags: [Organization]
      operationId: updatePosition
      summary: Edit or activate/deactivate a position definition
      parameters:
        - { name: positionId, in: path, required: true, schema: { type: string, format: uuid } }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/UpdatePosition' }
      responses:
        '200': { $ref: '#/components/responses/ObjectResponse' }
        '403': { $ref: '#/components/responses/Forbidden' }
    delete:
      tags: [Organization]
      operationId: deleteUnusedPosition
      summary: Delete an inactive and unreferenced position definition
      parameters:
        - { name: positionId, in: path, required: true, schema: { type: string, format: uuid } }
      responses:
        '204': { description: Deleted }
        '403': { $ref: '#/components/responses/Forbidden' }
  /api/v1/org/employees:
    get:
      tags: [Organization]
      operationId: listEmployees
      summary: List visible employees and active assignments
      responses:
        '200': { $ref: '#/components/responses/ListResponse' }
    post:
      tags: [Organization]
      operationId: createEmployee
      summary: Create an employee
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/CreateEmployee' }
      responses:
        '201': { $ref: '#/components/responses/ObjectResponse' }
  /api/v1/org/employees/{employeeId}:
    put:
      tags: [Organization]
      operationId: updateEmployee
      summary: Edit employee/account details or activate/deactivate the employee
      parameters:
        - $ref: '#/components/parameters/EmployeeId'
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/UpdateEmployee' }
      responses:
        '200': { $ref: '#/components/responses/ObjectResponse' }
        '403': { $ref: '#/components/responses/Forbidden' }
    delete:
      tags: [Organization]
      operationId: deleteUnusedEmployee
      summary: Delete an inactive employee without assignments, work or audit history
      parameters:
        - $ref: '#/components/parameters/EmployeeId'
      responses:
        '204': { description: Deleted }
        '403': { $ref: '#/components/responses/Forbidden' }
  /api/v1/org/employees/{employeeId}/assignments:
    post:
      tags: [Organization]
      operationId: createPositionAssignment
      summary: Assign a position to an employee; supports multiple positions
      parameters:
        - $ref: '#/components/parameters/EmployeeId'
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/CreatePositionAssignment' }
      responses:
        '201': { $ref: '#/components/responses/ObjectResponse' }
        '403': { $ref: '#/components/responses/Forbidden' }
  /api/v1/iam/me:
    get:
      tags: [IAM]
      operationId: getCurrentIdentity
      summary: Resolve the current account, roles, permissions, scopes and assignments
      responses:
        '200':
          description: Effective server-side identity
          content:
            application/json:
              schema: { $ref: '#/components/schemas/EffectiveIdentity' }
        '401': { $ref: '#/components/responses/Unauthorized' }
...
  /api/v1/iam/roles:
    get:
      tags: [IAM]
      operationId: listRoles
      summary: List tenant roles
      responses:
        '200': { $ref: '#/components/responses/ListResponse' }
    post:
      tags: [IAM]
      operationId: createRole
      summary: Create a tenant role
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/CreateRole' }
      responses:
        '201': { $ref: '#/components/responses/ObjectResponse' }
  /api/v1/iam/roles/{roleId}/permissions:
    put:
      tags: [IAM]
      operationId: replaceRolePermissions
      summary: Replace a role permission set
      parameters:
        - $ref: '#/components/parameters/RoleId'
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/SetRolePermissions' }
      responses:
        '204': { description: Permissions replaced }
  /api/v1/iam/role-assignments:
    post:
      tags: [IAM]
      operationId: grantRole
      summary: Grant a time-bounded role within an organization scope
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/GrantRole' }
      responses:
        '201': { $ref: '#/components/responses/ObjectResponse' }
...
  /api/v1/standard-evaluations:
    get:
      tags: [Evaluation]
      operationId: listStandardEvaluations
      summary: List standard evaluations
      parameters:
        - { name: outcome, in: query, schema: { type: string } }
        - { name: orgUnitId, in: query, schema: { type: string, format: uuid } }
      responses:
        '200': { $ref: '#/components/responses/ListResponse' }
    post:
      tags: [Evaluation]
      operationId: createStandardEvaluation
      summary: Create a deterministic standard evaluation idempotently
      description: >-
        Evaluates a submitted work record or a RESULT_SUBMITTED task against its bound,
        published standard version. Manual standard items leave the evaluation in
        PENDING_MANUAL for the manual-review endpoint.
      parameters:
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/CreateStandardEvaluation' }
      responses:
        '201': { $ref: '#/components/responses/ObjectResponse' }
        '400': { $ref: '#/components/responses/Problem' }
  /api/v1/standard-evaluations/{evaluationId}:
    get:
      tags: [Evaluation]
      operationId: getStandardEvaluation
      summary: Get evaluation summary and item results
      parameters:
        - $ref: '#/components/parameters/EvaluationId'
      responses:
        '200': { $ref: '#/components/responses/ObjectResponse' }
  /api/v1/standard-evaluations/{evaluationId}/manual-review:
    post:
      tags: [Evaluation]
      operationId: reviewStandardEvaluation
      summary: Complete manual evaluation items idempotently using expectedVersion
      parameters:
        - $ref: '#/components/parameters/EvaluationId'
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/ManualEvaluationReview' }
      responses:
        '200': { $ref: '#/components/responses/ObjectResponse' }
        '400': { $ref: '#/components/responses/Problem' }
  /api/v1/notifications:
    get:
      tags: [Notification]
      operationId: listNotifications
      summary: List notifications for the current account
      parameters:
        - name: unreadOnly
          in: query
          schema: { type: boolean, default: false }
      responses:
        '200': { $ref: '#/components/responses/ListResponse' }
  /api/v1/notifications/{notificationId}/read:
    post:
      tags: [Notification]
      operationId: markNotificationRead
      summary: Mark a notification as read using expectedVersion
      parameters:
        - $ref: '#/components/parameters/NotificationId'
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/MarkNotificationRead' }
      responses:
        '200': { $ref: '#/components/responses/ObjectResponse' }
        '409': { $ref: '#/components/responses/Conflict' }
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Trusted production JWT or SSO access token.
    TenantHeader:
      type: apiKey
      in: header
      name: X-Tenant-Id
      description: Local integration only; requires DEV_HEADER_AUTH_ENABLED=true.
    ActorHeader:
      type: apiKey
      in: header
      name: X-Actor-Id
      description: Local integration only; requires DEV_HEADER_AUTH_ENABLED=true.
  parameters:
    EmployeeId: { name: employeeId, in: path, required: true, schema: { type: string, format: uuid } }
    RoleId: { name: roleId, in: path, required: true, schema: { type: string, format: uuid } }
...
    EvaluationId: { name: evaluationId, in: path, required: true, schema: { type: string, format: uuid } }
    NotificationId: { name: notificationId, in: path, required: true, schema: { type: string, format: uuid } }
...
    IdempotencyKey:
      name: Idempotency-Key
      in: header
      required: true
      description: Client-generated key; reuse returns the original semantic result.
      schema: { type: string, minLength: 1, maxLength: 255 }
  responses:
    ListResponse:
      description: List response
      content:
        application/json:
          schema:
            type: array
            items: { type: object, additionalProperties: true }
    ObjectResponse:
      description: Object response
      content:
        application/json:
          schema: { type: object, additionalProperties: true }
    Problem:
      description: RFC 9457 problem detail
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/Problem' }
    Unauthorized:
      description: Missing, invalid or inactive identity
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/Problem' }
    Forbidden:
      description: Permission denied or organization scope violation
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/Problem' }
    Conflict:
      description: Optimistic version or idempotency conflict
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/Problem' }
  schemas:
    Problem:
      type: object
      properties:
        title: { type: string }
        status: { type: integer }
        detail: { type: string }
    EffectiveIdentity:
      type: object
      required: [tenantId, account, roles, permissions, tenantScope, organizationScopes, positionAssignments]
      properties:
        tenantId: { type: string, format: uuid }
        account: { type: object, additionalProperties: true }
        employee: { type: [object, 'null'], additionalProperties: true }
        primaryRole: { type: [string, 'null'] }
        roles: { type: array, uniqueItems: true, items: { type: string } }
        permissions: { type: array, uniqueItems: true, items: { type: string } }
        tenantScope: { type: boolean }
        organizationScopes: { type: array, uniqueItems: true, items: { type: string, format: uuid } }
        positionAssignments: { type: array, items: { type: object, additionalProperties: true } }
    CreateOrgUnit:
      type: object
      required: [code, name, unitType]
      properties:
        parentId: { type: [string, 'null'], format: uuid }
        code: { type: string }
        name: { type: string }
        unitType: { type: string, enum: [GROUP, REGION, HOTEL, DEPARTMENT] }
        sortOrder: { type: integer, default: 0 }
        propertyCode: { type: [string, 'null'] }
        city: { type: [string, 'null'] }
        roomCount: { type: [integer, 'null'], minimum: 0 }
        openingDate: { type: [string, 'null'], format: date }
    UpdateOrgUnit:
      type: object
      required: [code, name, status]
      properties:
        code: { type: string }
        name: { type: string }
        sortOrder: { type: integer, default: 0 }
        status: { type: string, enum: [ACTIVE, INACTIVE] }
        propertyCode: { type: [string, 'null'] }
        city: { type: [string, 'null'] }
        roomCount: { type: [integer, 'null'], minimum: 0 }
        openingDate: { type: [string, 'null'], format: date }
    CreatePosition:
      type: object
      required: [code, name, jobFamily]
      properties:
        code: { type: string }
        name: { type: string }
        jobFamily: { type: string }
        levelCode: { type: [string, 'null'] }
    UpdatePosition:
      type: object
      required: [code, name, jobFamily, status]
      properties:
        code: { type: string }
        name: { type: string }
        jobFamily: { type: string }
        levelCode: { type: [string, 'null'] }
        status: { type: string, enum: [ACTIVE, INACTIVE] }
    CreateEmployee:
      type: object
      required: [employeeNo, name]
      properties:
        employeeNo: { type: string }
        name: { type: string }
        mobile: { type: [string, 'null'] }
        hiredOn: { type: [string, 'null'], format: date }
        loginName: { type: [string, 'null'] }
        temporaryPassword: { type: [string, 'null'], format: password, minLength: 10 }
    UpdateEmployee:
      type: object
      required: [employeeNo, name, employmentStatus]
      properties:
        employeeNo: { type: string }
        name: { type: string }
        mobile: { type: [string, 'null'] }
        hiredOn: { type: [string, 'null'], format: date }
        employmentStatus: { type: string, enum: [ACTIVE, INACTIVE] }
        loginName: { type: [string, 'null'] }
        temporaryPassword: { type: [string, 'null'], format: password, minLength: 10 }
    CreatePositionAssignment:
      type: object
      required: [orgUnitId, positionId, validFrom]
      properties:
        orgUnitId: { type: string, format: uuid }
        positionId: { type: string, format: uuid }
        managerAssignmentId: { type: [string, 'null'], format: uuid }
        primary: { type: boolean, default: false }
        assignmentType: { type: string, enum: [PERMANENT, TEMPORARY, ACTING] }
        validFrom: { type: string, format: date }
        validTo: { type: [string, 'null'], format: date }
    CreateRole:
      type: object
      required: [code, name]
      properties:
        code: { type: string }
        name: { type: string }
        roleType: { type: [string, 'null'] }
    SetRolePermissions:
      type: object
      required: [permissionIds]
      properties:
        permissionIds:
          type: array
          minItems: 1
          items: { type: string, format: uuid }
    GrantRole:
      type: object
      required: [accountId, roleId, scopeType]
      properties:
        accountId: { type: string, format: uuid }
        roleId: { type: string, format: uuid }
        scopeOrgUnitId: { type: [string, 'null'], format: uuid }
        scopeType: { type: string, enum: [SELF, ORG_UNIT, ORG_TREE, TENANT] }
        validFrom: { type: [string, 'null'], format: date-time }
        validTo: { type: [string, 'null'], format: date-time }
...
    CreateStandardEvaluation:
      type: object
      required: [subjectType, subjectId, orgUnitId, standardVersionId, inputSnapshot]
      properties:
        subjectType: { type: string, enum: [WORK_RECORD, TASK] }
        subjectId: { type: string, format: uuid }
        orgUnitId: { type: string, format: uuid }
        positionAssignmentId: { type: [string, 'null'], format: uuid }
        standardVersionId: { type: string, format: uuid }
        inputSnapshot: { type: object, additionalProperties: true }
    ManualEvaluationItem:
      type: object
      required: [itemCode, outcome]
      properties:
        itemCode: { type: string }
        outcome: { type: string, enum: [PASS, WARNING, FAIL] }
        score: { type: [number, 'null'] }
        reason: { type: [string, 'null'] }
        actualValue: {}
    ManualEvaluationReview:
      type: object
      required: [expectedVersion, reviewerAssignmentId, items]
      properties:
        expectedVersion: { type: integer, format: int64, minimum: 0 }
        reviewerAssignmentId: { type: string, format: uuid }
        items:
          type: array
          minItems: 1
          items: { $ref: '#/components/schemas/ManualEvaluationItem' }
    MarkNotificationRead:
      type: object
      required: [expectedVersion]
      properties:
        expectedVersion: { type: integer, format: int64, minimum: 0 }
...